% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverMargins.R
\name{CoverMargins}
\alias{CoverMargins}
\title{Extract a set of model terms that covers a formula}
\usage{
CoverMargins(data, formula, coverTable = FALSE)
}
\arguments{
\item{data}{Input data as a data frame (inner cells)}

\item{formula}{Model formula defining cells to consider}

\item{coverTable}{Whether to return a model term defining a cover table}
}
\value{
Vector of extracted model terms or a two-component list when coverTable is TRUE
}
\description{
Other model terms depend on the extracted set. Hierarchical relationships seen in the data is considered.
}
\note{
all terms retuned when coverTable=NA
}
\examples{

z3 <- EasyData("z3")
CoverMargins(z3, ~region + kostragr * hovedint)
CoverMargins(z3, ~region * mnd + hovedint * mnd + fylke * hovedint * mnd + 
                  kostragr * hovedint * mnd)
CoverMargins(z3, ~fylke * hovedint + kostragr * hovedint, TRUE)
CoverMargins(z3, ~hovedint * mnd2 + mnd + hovedint + kostragr * hovedint, TRUE)
}
