% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverMargins.R
\name{CoverMarginsExact}
\alias{CoverMarginsExact}
\title{Extract two sets of model terms that cover two formulas}
\usage{
CoverMarginsExact(data, formula, formulaExact = NULL,
  intercept = FALSE, interceptExact = !is.null(formulaExact),
  coverMargins = TRUE)
}
\arguments{
\item{data}{Input data as a data frame (inner cells)}

\item{formula}{Model formula defining cells to be perturbed}

\item{formulaExact}{Model formula defining cells to be exact}

\item{intercept}{Whether to include intercept}

\item{interceptExact}{Whether to include intercept}

\item{coverMargins}{Whether to extract only the covering terms}
}
\value{
List of two sets of model terms
}
\description{
One formula for cells to be perturbed and one formula for cells to be exact
}
\details{
The function make use of \code{\link{CoverMargins}}. Thus, hierarchical relationships seen in the data is considered. 
Each output element can be input to \code{\link{Terms2formula}}.
}
\examples{
z3 <- EasyData("z3")
CoverMarginsExact(z3, ~region + kostragr * hovedint)
CoverMarginsExact(z3, ~region + kostragr * hovedint * mnd, ~kostragr * hovedint)
CoverMarginsExact(z3, ~region + kostragr * hovedint * mnd, ~kostragr * hovedint + mnd)
CoverMarginsExact(z3, ~region, ~kostragr * hovedint, TRUE, FALSE)
CoverMarginsExact(z3, ~region * mnd, ~kostragr * hovedint, coverMargins = FALSE)
CoverMarginsExact(z3, ~region, interceptExact = TRUE)  
}
\author{
Øyvind Langsrud
}
