% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{Lap}
\alias{Lap}
\title{Laplace noise}
\usage{
Lap(n = 1, b = 1)
}
\arguments{
\item{n}{number of observations}

\item{b}{scale parameter(s)}
}
\value{
random deviates
}
\description{
Laplace noise
}
\examples{
Lap(4)
Lap(4, 20)
Lap(4, 1:4)
}
