% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverMargins.R
\name{NoisyCoverMargins}
\alias{NoisyCoverMargins}
\title{Laplace or pTable perturbed margins}
\usage{
NoisyCoverMargins(data, freqVar, formula, formulaExact = NULL,
  eps = 0.5, pMatrix = NULL, rkeys = NULL, ...)
}
\arguments{
\item{data}{Input data as a data frame (inner cells)}

\item{freqVar}{Variable holding counts (name or number)}

\item{formula}{Model formula defining cells to be perturbed}

\item{formulaExact}{Model formula defining cells to be exact}

\item{eps}{Differential privacy parameter}

\item{pMatrix}{Output from \code{\link{Pmatrix}}}

\item{rkeys}{Uniformly distributed keys}

\item{...}{Further parameters to \code{\link{CoverMarginsExact}}}
}
\value{
A list
}
\description{
The function make use of \code{\link{CoverMarginsExact}}.
}
\examples{
z2 <- EasyData("z2")
a1 <- NoisyCoverMargins(z2, "ant", ~region + kostragr * hovedint)
a2 <- NoisyCoverMargins(z2, "ant", ~region + kostragr * hovedint, ~kostragr + hovedint)
}
