% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce.R
\name{ReduceBy0}
\alias{ReduceBy0}
\title{Reduce input to \code{\link{Mipf}} based on zeros in z and possible zeros in yStart}
\usage{
ReduceBy0(x, z, yStart = NULL)
}
\arguments{
\item{x}{a matrix}

\item{z}{a single column matrix}

\item{yStart}{a starting estimate of \code{y}}
}
\value{
A list of three elements:
        \item{\code{x}}{Reduced \code{x}}
        \item{\code{z}}{Corresponding reduced \code{z}}
        \item{\code{yKnown}}{Logical vector specifying elements of y that can be found directly as 0}
}
\description{
Reduce input to \code{\link{Mipf}} based on zeros in z and possible zeros in yStart
}
\examples{
# Generate data with zeros
z2 <- EasyData("z2")
z2$ant[z2$hovedint == "trygd"] <- 0
z2$ant[1:3] <- 0

x <- FormulaSums(z2, ~~region + kostragr * hovedint - 1)
z <- t(x) \%*\% z2$ant

# Run ReduceBy0
a <- ReduceBy0(x, z)

# The rows known to be 0 and omitted from a$x
z2[a$yKnown, ]

# Dimension of x reduced
dim(x)
dim(a$x)

# Additional elements know to be 0
b <- ReduceBy0(x, z, c(0, 0, rep(1, 42)))
z2[b$yKnown, ]
}
\author{
Øyvind Langsrud
}
\keyword{internal}
