% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReleasableInner.R
\name{ReleasableInner}
\alias{ReleasableInner}
\alias{NoisyCoverMarginsToInner}
\title{Releasable inner cell frequencies by post-processing protected tabular data}
\usage{
ReleasableInner(
  data,
  freqVar,
  formula,
  formulaExact = NULL,
  regMethod = c("nnls", "ginv", "glmnet"),
  returnDetails = FALSE,
  rerun = 0,
  eps = 0.5,
  pMatrix = NULL,
  keyVar = NULL,
  wExact = 1000,
  intercept = FALSE,
  interceptExact = !is.null(formulaExact),
  coverMargins = TRUE,
  ...
)

NoisyCoverMarginsToInner(
  obj,
  regMethod = c("nnls", "ginv", "glmnet"),
  returnDetails = FALSE,
  rerun = 0,
  ...
)
}
\arguments{
\item{data}{Input data as a data frame (inner cells)}

\item{freqVar}{Variable holding counts (name or number)}

\item{formula}{Model formula defining cells (margins) to be perturbed}

\item{formulaExact}{Model formula defining cells (margins) to be exact}

\item{regMethod}{One of "nnls" (default),"ginv" or"glmnet". See \code{\link{WRRnnls}}.}

\item{returnDetails}{Whether to include all output from the underlying functions.}

\item{rerun}{Possibility for a variant that iterates. Ordinary margins are re-computed in each round while the exact margins are unchanged.}

\item{eps}{Differential privacy parameter}

\item{pMatrix}{Output from \code{\link{Pmatrix}}}

\item{keyVar}{Variable holding uniformly distributed keys (name or number).}

\item{wExact}{Weight for highly weighted margins}

\item{intercept}{Parameter to \code{\link{CoverMarginsExact}}}

\item{interceptExact}{Parameter to \code{\link{CoverMarginsExact}}}

\item{coverMargins}{Parameter to \code{\link{CoverMarginsExact}}}

\item{...}{Parameters to or \code{\link{WRRnnls}}, \code{\link{WRRginv}} or \code{\link{WRRglmnet}}}

\item{obj}{Output from \code{\link{NoisyCoverMargins}}}
}
\value{
Estimated inner cells frequencies (yHat) or a list including all output from
        \code{\link{NoisyCoverMargins}} and \code{\link{WRRnnls}}, \code{\link{WRRginv}} or \code{\link{WRRglmnet}}.
}
\description{
Margins are first perturbed according to pTable or differential privacy (Laplace noise) by
\code{\link{NoisyCoverMargins}}.
Thereafter inner cells are estimated from the perturbed margins by
\code{\link{WRRnnls}}, \code{\link{WRRginv}} or \code{\link{WRRglmnet}}.
}
\examples{
y <- c(3, 1, 5, 6, 4, 8, 2, 7, 27)
cols <- paste("col", col(matrix(1:9, 3, 3)), sep = "")
rows <- paste("row", row(matrix(1:9, 3, 3)), sep = "")
z <- data.frame(rows, cols, y)
pMatrix <- Pmatrix()
z$keys <- runif(9)

# Non-negative estimates without noise
ReleasableInner(z, "y", ~rows + cols, eps = Inf)

# Non-negative estimates with Laplace noise
ReleasableInner(z, "y", ~rows + cols, eps = 0.5)

# Using pTable and keys. Overall total perturbed.
ReleasableInner(z, "y", ~rows + cols, pMatrix = pMatrix, keyVar = "keys", intercept = TRUE)

# Generalized inverse (not non-negative)
ReleasableInner(z, "y", ~rows + cols, pMatrix = pMatrix, keyVar = "keys", intercept = TRUE, 
                regMethod = "ginv")

# Preserving row totals
ReleasableInner(z, "y", ~rows + cols, ~rows)

# Other data and detailed output
z2 <- EasyData("z2")
a1 <- ReleasableInner(z2, "ant", ~region + kostragr * hovedint, ~kostragr + hovedint, 
                      returnDetails = TRUE)

# Other values of thresh and wExact
a2 <- ReleasableInner(z2, "ant", ~region + kostragr * hovedint, ~kostragr + hovedint, 
                      regMethod = "glmnet", thresh = 1E-8, wExact = 100)

# Not non-negative
a3 <- ReleasableInner(z2, "ant", ~region + kostragr * hovedint, ~kostragr + hovedint, 
                      regMethod = "glmnet", lower.limits = -Inf)
}
\seealso{
Another way to obtain releasable inner cell frequencies from protected tabular data is to 
          run \code{\link{SuppressDec}} with input from \code{\link{PTxyz}}.
}
\author{
Øyvind Langsrud
}
