% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{Terms2formula}
\alias{Terms2formula}
\title{Model formula from vector of terms where intercept is specified as \code{"1"}}
\usage{
Terms2formula(x, InterceptAs1 = TRUE, env = parent.frame())
}
\arguments{
\item{x}{Character vector}

\item{InterceptAs1}{Whether to specify intercept as \code{"1"}}

\item{env}{Parameter to \code{\link{as.formula}}}
}
\value{
formula
}
\description{
Model formula from vector of terms where intercept is specified as \code{"1"}
}
\examples{
Terms2formula(c("a", "b:c"))
Terms2formula(c("a", "b:c"), FALSE)
Terms2formula(c("a", "b:c", "1"))
}
