% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedRidgeRegression.R
\name{WRRnnls}
\alias{WRRnnls}
\alias{WRRginv}
\alias{WRRglmnet}
\title{Special weighted and ridge penalized regression}
\usage{
WRRnnls(x, y, xExact = NULL, yExact = NULL, wExact = 1000,
  lambda = 1e-04^2)

WRRginv(x, y, xExact = NULL, yExact = NULL, wExact = 1000,
  lambda = 1e-04^2)

WRRglmnet(x, y, xExact = NULL, yExact = NULL, wExact = 1000,
  lambda = exp(((2 * 12):(-17 * 2))/2), intercept = FALSE,
  standardize = FALSE, thresh = 1e-10, lower.limits = 0, ...)
}
\arguments{
\item{x}{Input matrix, each row is an ordinary observation}

\item{y}{Ordinary response observation (vector or matrix)}

\item{xExact}{Input matrix, each row is a highly weighted observation}

\item{yExact}{Highly weighted response observation (vector or matrix)}

\item{wExact}{Weight for highly weighted observations}

\item{lambda}{Ridge regression penalty parameter (sequence when glmnet)}

\item{intercept}{glmnet parameter}

\item{standardize}{glmnet parameter}

\item{thresh}{glmnet parameter}

\item{lower.limits}{glmnet parameter}

\item{...}{Further glmnet parameters}
}
\value{
Output from \code{\link{nnls}}, \code{\link{glmnet}} or coefficient estimate calculated using \code{\link{ginv}}
}
\description{
By using \code{\link{nnls}}, \code{\link{ginv}} or \code{\link{glmnet}}
}
\examples{
x <- cbind(1:11, -11:-1, c(1, 1, 2, 2, 1, 1, 2, 2, 1, 1, 2))
y <- matrix(5 - sin(2 * (1:11)), dimnames = list(NULL, "y"))
x1 <- x[1:9, ]
x2 <- x[10:11, ]
y1 <- y[1:9, , drop = FALSE]
y2 <- y[10:11, , drop = FALSE]


# Generalized inverse estimation
ginvCoef <- WRRginv(x1, y1, x2, y2)

# Non-negative estimation by nnls 
# Coefficients in output element x 
nnlsCoef <- WRRnnls(x1, y1, x2, y2)$x

# Non-negative estimation by glmnet
# Take out best fit from matrix of coefficients 
gn <- WRRglmnet(x1, y1, x2, y2)
glmnetCoef <- coef(gn)[-1, which.max(gn$dev.ratio), drop = FALSE]

# Another estimation by glmnet (not non-negative)
# Take out best fit from matrix of coefficients
gnInf <- WRRglmnet(x1, y1, x2, y2, lower.limits = -Inf)
glmnetCoefInf <- coef(gnInf)[-1, which.max(gn$dev.ratio), drop = FALSE]

# All coefficients
coef4 <- as.matrix(cbind(ginvCoef, nnlsCoef, glmnetCoef, glmnetCoefInf))
colnames(coef4) <- c("ginv", "nnls", "glmnet", "glmnetInf")
print(coef4)

# Original y and fitted values. Close fit for last two observation.
cbind(y, x \%*\% coef4)
}
\author{
Øyvind Langsrud
}
