% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverMargins.R
\name{NoisyCoverMargins}
\alias{NoisyCoverMargins}
\title{Laplace or pTable perturbed margins}
\usage{
NoisyCoverMargins(data, freqVar, formula, formulaExact = NULL,
  eps = 0.5, pMatrix = NULL, keyVar = NULL, ...)
}
\arguments{
\item{data}{Input data as a data frame (inner cells)}

\item{freqVar}{Variable holding counts (name or number)}

\item{formula}{Model formula defining cells to be perturbed}

\item{formulaExact}{Model formula defining cells to be exact}

\item{eps}{Differential privacy parameter}

\item{pMatrix}{Output from \code{\link{Pmatrix}}}

\item{keyVar}{Variable holding uniformly distributed keys (name or number). See details.}

\item{...}{Further parameters to \code{\link{CoverMarginsExact}}}
}
\value{
A list
}
\description{
The function make use of \code{\link{CoverMarginsExact}}.
}
\details{
In this function individual keys (cell keys) are not taken as input. 
Possible individual keys must be "summed up" to inner cell keys beforehand. 
Zeros keys corresponding to zero frequencies are unproblematic when zeros are not perturbed. 
By default (\code{keyVar==NULL}), the keys corresponding to nonzero frequencies are generated by runif (otherwise zero).
}
\examples{
z2 <- EasyData("z2")
a1 <- NoisyCoverMargins(z2, "ant", ~region + kostragr * hovedint)
a2 <- NoisyCoverMargins(z2, "ant", ~region + kostragr * hovedint, ~kostragr + hovedint, eps = 0.25)
a3 <- NoisyCoverMargins(cbind(z2,keyVar = runif(44) * (z2$ant>0)), "ant", 
                        ~region + kostragr * hovedint, ~kostragr + hovedint, 
                        pMatrix = Pmatrix(), keyVar = "keyVar")
}
\author{
Øyvind Langsrud
}
