% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{Terms2formula}
\alias{Terms2formula}
\title{Model formula from vector of terms where intercept is specified as a string}
\usage{
Terms2formula(x, intercept = "(Intercept)", env = parent.frame())
}
\arguments{
\item{x}{Character vector}

\item{intercept}{String to specify intercept}

\item{env}{Parameter to \code{\link{as.formula}}}
}
\value{
formula
}
\description{
Model formula from vector of terms where intercept is specified as a string
}
\examples{
Terms2formula(c("a", "b:c"))
Terms2formula(c("a", "b:c"), NULL)
Terms2formula(c("a", "b:c", "(Intercept)"))
Terms2formula(c("a", "b:c"), "1")
Terms2formula(c("a", "b:c", "1"), "1")
}
